chore: move sinks tests to internal tests folder
