+++ platform_plugin_aspects/signals.py
@@ -7,7 +7,7 @@ from django.dispatch import Signal, receiver
 
 from platform_plugin_aspects.sinks.external_id_sink import ExternalIdSink
 from platform_plugin_aspects.sinks.user_profile_sink import UserProfileSink
-from platform_plugin_aspects.sinks.user_retire import UserRetirementSink
+from platform_plugin_aspects.sinks.user_retire_sink import UserRetirementSink
 from platform_plugin_aspects.utils import get_model
 
 try:
