+++ platform_plugin_aspects/sinks/course_overview_sink.py
